[Refactor] Move flat JSON column transforms to ColumnCore
