+++ be/src/column/flat_json/flat_json_internal.h
@@ -21,9 +21,9 @@
 #include <string_view>
 
 #include "base/phmap/phmap.h"
+#include "column/flat_json/json_flat_path.h"
 #include "column/nullable_column.h"
 #include "types/logical_type.h"
-#include "util/json_flat_path.h"
 
 namespace starrocks {
