+++ be/src/storage/rowset/json_column_compactor.cpp
@@ -21,6 +21,9 @@
 #include <vector>
 
 #include "column/column.h"
+#include "column/flat_json/json_flat_path.h"
+#include "column/flat_json/json_flattener.h"
+#include "column/flat_json/json_merger.h"
 #include "column/json_column.h"
 #include "column/nullable_column.h"
 #include "gen_cpp/segment.pb.h"
@@ -29,9 +32,6 @@
 #include "storage/rowset/column_writer.h"
 #include "types/constexpr.h"
 #include "util/hyper_json_transformer.h"
-#include "util/json_flat_path.h"
-#include "util/json_flattener.h"
-#include "util/json_merger.h"
 #include "util/json_path_deriver.h"
 
 namespace starrocks {
