+++ be/src/storage/rowset/segment.cpp
@@ -45,6 +45,7 @@
 #include "base/string/slice.h"
 #include "base/utility/defer_op.h"
 #include "column/column_access_path.h"
+#include "column/flat_json/json_flat_path.h"
 #include "column/schema.h"
 #include "common/bloom_filter.h"
 #include "common/config_rowset_fwd.h"
@@ -67,7 +68,6 @@
 #include "storage/storage_metrics.h"
 #include "storage/tablet_schema.h"
 #include "storage/utils.h"
-#include "util/json_flat_path.h"
 
 bvar::Adder<int> g_open_segments;    // NOLINT
 bvar::Adder<int> g_open_segments_io; // NOLINT
