+++ be/src/util/hyper_json_transformer.cpp
@@ -23,6 +23,8 @@
 
 #include "column/column_helper.h"
 #include "column/const_column.h"
+#include "column/flat_json/json_flattener.h"
+#include "column/flat_json/json_merger.h"
 #include "column/json_column.h"
 #include "column/nullable_column.h"
 #include "common/runtime_profile.h"
@@ -31,8 +33,6 @@
 #include "exprs/expr_context.h"
 #include "runtime/descriptors.h"
 #include "types/type_descriptor.h"
-#include "util/json_flattener.h"
-#include "util/json_merger.h"
 
 namespace starrocks {
