+++ be/test/exprs/flat_json_functions_test.cpp
@@ -24,6 +24,7 @@
 #include "base/utility/defer_op.h"
 #include "butil/time.h"
 #include "column/const_column.h"
+#include "column/flat_json/json_flattener.h"
 #include "column/map_column.h"
 #include "column/nullable_column.h"
 #include "column/struct_column.h"
@@ -38,7 +39,6 @@
 #include "gutil/strings/strip.h"
 #include "types/json_value.h"
 #include "types/logical_type.h"
-#include "util/json_flattener.h"
 #include "util/json_path_deriver.h"
 
 namespace starrocks {
